Use size_t instead

+++ src/fheroes2/gui/interface_events.cpp
@@ -138,8 +138,8 @@ void Interface::Basic::EventNextHero( void )
         } while ( it != currentHero );
     }
     else {
-        const int heroesCount = myHeroes.size();
-        for ( int i = 0; i < heroesCount; ++i ) {
+        const size_t heroesCount = myHeroes.size();
+        for ( size_t i = 0; i < heroesCount; ++i ) {
             if ( myHeroes[i]->MayStillMove() ) {
                 SetFocus( myHeroes[i] );
                 break;
